server/scripts/python/requirements.txt
marker-pdf>=0.3.2
python-dotenv>=1.0.0
Pillow>=10.0.0
